app/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
boto3==1.34.160
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.12
Pillow==10.3.0
pillow-heif==0.16.0
cairosvg==2.7.1
opencv-python-headless==4.10.0.84
# VLM stack
torch==2.3.1
transformers==4.45.2
accelerate==0.33.0
sentencepiece==0.2.0
huggingface_hub==0.24.5 bitsandbytes==0.44.1
flash-attn==2.6.3
websockets==12.0
redis==5.1.1
